CEA_gscholar_scrap/requirements.txt
fastapi==0.96.0
google-serp-api==1.0.3
scholarly==1.7.11
uvicorn==0.22.0
google-search-results==2.4.2
pymongo==4.3.3